+++ merge_request_approvals_test.go
@@ -364,9 +364,9 @@ func TestCreateApprovalRules(t *testing.T) {
 
 	opt := &CreateMergeRequestApprovalRuleOptions{
 		Name:              Ptr("security"),
-		ApprovalsRequired: Ptr(3),
-		UserIDs:           &[]int{5, 50},
-		GroupIDs:          &[]int{5},
+		ApprovalsRequired: Ptr(int64(3)),
+		UserIDs:           &[]int64{5, 50},
+		GroupIDs:          &[]int64{5},
 	}
 
 	rule, resp, err := client.MergeRequestApprovals.CreateApprovalRule(1, 1, opt)
